+++ commands/poll.js
@@ -26,7 +26,8 @@ const log = require('log4js').getLogger('amy');
 const { createQuickPoll, createGamePoll } = require('../lib/Survey');
 
 // poll quick [question]
-// poll game [times] [questions]
+// poll game [times,] [game]
+// poll mc [options,];[question]
 
 // Handler for poll command
 module.exports = async (bot, msg, args) => {
@@ -75,7 +76,41 @@ module.exports = async (bot, msg, args) => {
 }
 
 function createExamples(channel, quick = false, game = false, multipleChoice = false) {
+    let catdance = "<a:catdance:729440916128923649>";
+    let example = {
+        "content": "<a:ditto:729440915835322379> So you want some **polling** examples! <a:ditto:729440915835322379>",
+        "embed": {
+            "title": "We can make cool polls, yes.",
+            "description": "Here are some examples.",
+            "fields": [],
+            "footer": {
+                "text": "!help always calls the help command."
+            }
+        }
+    };
     if (!(quick && game && multipleChoice)) {
         quick = true, game = true, multipleChoice = true;
+    }
+    if (quick) {
+        let quickInfo = {
+            "name": "Quick Polls",
+            "value": "You can make quick polls to ask your friends yes/no questions.\n\nTry asking your friends questions with commands like:\n"
+        };
+    }
+    if (game) {
+        let gameInfo = {
+            "name": "GamePolls",
+            "value": "You can make game polls to see if anyone's down to play a game.\n\nTry making or with commands like:\n"
+        };
+        let timeInfo = {
+            "name": "Formatting Time",
+            "value": "Separate the times you want with commas, and no spaces in between. If you put spaces, Amy will try to figure out what you meant, but she isn't *that* smart yet.\n\n<a:catdance:729440916128923649> `3,4,5` means **3 PM, 4 PM, 5 PM**\n<a:catdance:729440916128923649> `3h,4h,5h` means **in 3 hours, in 4 hours, in 5 hours**. You can also use `m` or `d` for months or days. *It'll automatically be converted into an approximate date and time.*\n<a:catdance:729440916128923649> `3:30,4:30,5:30` means **3:30 PM, 4:30 PM, 5:30 PM**\n\nYou're  also allowed to mix and match, so `3,4:15,8h` is **3 PM, 4:15 PM, in 8 hours**."
+        };
+    }
+    if (multipleChoice) {
+        let mcInfo = {
+            "name": "Multiple Choice Polls",
+            "value": "You can make multiple choice polls for additional flexibility. List your options, then a semicolon, then the question, like so:\n\n"
+        };
     }
 }